+++ index_binary.go
@@ -107,7 +107,7 @@ type BinaryIndex interface {
 	bPtr() *C.FaissIndexBinary
 
 	// CodeSize returns the size of the produced codes in bytes.
-	CodeSize() (int, error)
+	CodeSize() (uint64, error)
 }
 
 type faissBinaryIndex struct {
@@ -424,12 +424,12 @@ func (b *faissBinaryIndex) Size() uint64 {
 	return rv
 }
 
-func (b *faissBinaryIndex) CodeSize() (int, error) {
+func (b *faissBinaryIndex) CodeSize() (uint64, error) {
 	var size C.size_t
 	if c := C.faiss_IndexBinary_code_size(b.bIdx, &size); c != 0 {
 		return 0, getLastError()
 	}
-	return int(size), nil
+	return uint64(size), nil
 }
 
 func (idx *faissBinaryIndex) Close() {
